fix: Update import paths after component reorganization
-  Fixed import paths in all page files to match new component structure
-  Updated imports to use organized directories (auth/, common/, dashboard/, pages/)
-  Resolved 'Module not found' build errors
-  All pages now correctly import from reorganized component locations

Pages updated:
- app/page.tsx (main landing page)
- app/login/page.tsx
- app/signup/page.tsx
- app/admin/page.tsx
- app/student/page.tsx
- app/mentor/page.tsx

Build now compiles successfully without module resolution errors.
