+++ Frontend/app/admin/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import Admin from "../components/Admin";
-import Navbar from "../components/Navbar";
-import Footer from "../components/Footer";
+import Admin from "../components/pages/Admin";
+import Navbar from "../components/common/Navbar";
+import Footer from "../components/common/Footer";
 
 export default function AdminPage() {
   return (
