+++ Frontend/app/mentor/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import MentorDashboard from "../components/MentorDashboard";
-import Navbar from "../components/Navbar";
-import Footer from "../components/Footer";
+import MentorDashboard from "../components/dashboard/MentorDashboard";
+import Navbar from "../components/common/Navbar";
+import Footer from "../components/common/Footer";
 
 export default function MentorPage() {
   return (
