+++ Frontend/app/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import React, { useState, useEffect } from "react";
-import Navbar from "./components/Navbar";
-import Footer from "./components/Footer";
-import AnimatedElement from "./components/AnimatedElement";
+import Navbar from "./components/common/Navbar";
+import Footer from "./components/common/Footer";
+import AnimatedElement from "./components/common/AnimatedElement";
 
 const navLinks = [
   { label: "Home", href: "/" },
