+++ Frontend/app/signup/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import Signup from "../components/Signup";
-import Navbar from "../components/Navbar";
-import Footer from "../components/Footer";
+import Signup from "../components/auth/Signup";
+import Navbar from "../components/common/Navbar";
+import Footer from "../components/common/Footer";
 
 export default function SignupPage() {
   return (
