+++ Frontend/app/student/page.tsx
@@ -1,7 +1,7 @@
 "use client";
-import StudentDashboard from "../components/StudentDashboard";
-import Navbar from "../components/Navbar";
-import Footer from "../components/Footer";
+import StudentDashboard from "../components/dashboard/StudentDashboard";
+import Navbar from "../components/common/Navbar";
+import Footer from "../components/common/Footer";
 
 export default function StudentPage() {
   return (
